requirements.txt
python-slugify
wand

